test(countup): remove unpredictable test
